+++ gapipy/resources/booking/__init__.py
@@ -18,6 +18,8 @@ from .service import (
     FlightService,
     InsuranceService,
     MerchandiseService,
+    RegionalConnectorService,
+    RoomUpgradeService,
     SingleSupplementService,
     TransportService,
 )